Adding scaled dot product attention layer
